third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamTrack-getCapabilities.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio: true})
      .then(function(stream) {
    var capabilities = stream.getAudioTracks()[0].getCapabilities();
    assert_equals(Object.keys(capabilities).length, 2);
    assert_true(capabilities.hasOwnProperty('deviceId'));
    assert_true(capabilities.hasOwnProperty('echoCancellation'));
    assert_equals(Object.keys(capabilities.echoCancellation).length, 2);
  });
}, 'getCapabilities() support for getUserMedia() audio track.');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({video: true})
      .then(function(stream) {
    var capabilities = stream.getVideoTracks()[0].getCapabilities();
    assert_greater_than(Object.keys(capabilities).length, 0);
    assert_true(capabilities.hasOwnProperty('deviceId'));
    if (capabilities.hasOwnProperty('width')) {
      assert_equals(Object.keys(capabilities.width).length, 2);
      assert_true(capabilities.width.hasOwnProperty('min'));
      assert_true(capabilities.width.hasOwnProperty('max'));
      assert_less_than_equal(capabilities.width.min, capabilities.width.max);
    }
    if (capabilities.hasOwnProperty('height')) {
      assert_equals(Object.keys(capabilities.height).length, 2);
      assert_true(capabilities.height.hasOwnProperty('min'));
      assert_true(capabilities.height.hasOwnProperty('max'));
      assert_less_than_equal(capabilities.height.min, capabilities.height.max);
    }
    if (capabilities.hasOwnProperty('aspectRatio')) {
      assert_equals(Object.keys(capabilities.aspectRatio).length, 2);
      assert_true(capabilities.aspectRatio.hasOwnProperty('min'));
      assert_true(capabilities.aspectRatio.hasOwnProperty('max'));
      assert_less_than_equal(capabilities.aspectRatio.min, capabilities.aspectRatio.max);
    }
    if (capabilities.hasOwnProperty('frameRate')) {
      assert_equals(Object.keys(capabilities.frameRate).length, 2);
      assert_true(capabilities.frameRate.hasOwnProperty('min'));
      assert_true(capabilities.frameRate.hasOwnProperty('max'));
      assert_less_than_equal(capabilities.frameRate.min, capabilities.frameRate.max);
    }
    assert_true(capabilities.hasOwnProperty('facingMode'));
  });
}, 'getCapabilities() support for getUserMedia() video track.');
</script>
